requirements.txt
# Core
tqdm>=4.66.0
numpy>=1.26.0
pandas>=2.2.0

# Vector DB
faiss-cpu>=1.8.0

# Data / PDFs
datasets>=2.20.0
pypdf>=4.2.0
pdfplumber>=0.11.0
python-docx
# App
streamlit>=1.35.0

# Utils
python-dotenv>=1.0.1
tiktoken>=0.7.0

jupyterlab>=4.2.0
ipykernel>=6.29.0